add support for exceptions
This adds support for exceptions which are applied both on write to
Tigerblood, and on read. Submitting an address in the exception list
will cause the address to be effectively ignored and not tracked, and
likewise reading a reputation for something in the exception list will
cause it to appear as untracked.

The initial implementation supports static exceptions which persist for
the lifetime of the process read from flat files, and dynamic exceptions
which are periodically updated. Current support for dynamic exceptions
includes polling known AWS public IP address ranges.