add new reviewed value for reputation entry
Intent is to support explicitly marking a given reputation entry in the
database as verified/reviewed.

Defaults to false, but can be set to true in API insert/update
commands for reputation entries.

When an entry is updated with a reputation score of 100, the review flag
is toggled to false again.

Closes #97